+++ src/main.rs
@@ -1,13 +1,12 @@
 mod cli;
 mod error;
-
-use std::collections::HashMap;
+mod ssr;
 
 use crate::cli::Cli;
 use crate::error::{Error, Result};
+use crate::ssr::SsrRecord;
 use cli::Environment;
 use reqwest::blocking::Client;
-use serde::{Deserialize, Serialize};
 
 fn env_parse(targets: &Vec<Environment>) -> Vec<(String, String)> {
     if targets.len() == 0 {
@@ -25,54 +24,17 @@ fn env_parse(targets: &Vec<Environment>) -> Vec<(String, String)> {
     results
 }
 
-#[derive(Debug, Serialize, Deserialize)]
-struct Ssr {
-    name: String,
-    description: String,
-    key: String,
-    url: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct SsrResult {
-    name: String,
-    description: String,
-    key: String,
-    url: HashMap<String, String>,
-}
-
-impl SsrResult {
-    pub fn new(
-        name: impl Into<String>,
-        description: impl Into<String>,
-        key: impl Into<String>,
-    ) -> Self {
-        SsrResult {
-            name: name.into(),
-            description: description.into(),
-            key: key.into(),
-            url: HashMap::new(),
-        }
-    }
-}
-
-impl SsrResult {
-    pub fn update_url(&mut self, target: impl Into<String>, value: impl Into<String>) {
-        let _ = self.url.insert(target.into(), value.into());
-    }
-}
-
 fn get_records(
     client: &reqwest::blocking::RequestBuilder,
     target: &(String, String),
     pattern: &Option<String>,
-) -> Result<Vec<Ssr>> {
+) -> Result<Vec<SsrRecord>> {
     let result = client
         .try_clone()
         .ok_or_else(|| Error::UnableToCloneClient)?
         .query(&[target])
         .send()?
-        .json::<Vec<Ssr>>()?
+        .json::<Vec<SsrRecord>>()?
         .into_iter()
         .filter(|record| match &pattern {
             Some(value) => {
@@ -82,7 +44,7 @@ fn get_records(
             }
             None => true,
         })
-        .collect::<Vec<Ssr>>();
+        .collect::<Vec<SsrRecord>>();
     Ok(result)
 }
 
@@ -92,7 +54,7 @@ fn main() -> Result<()> {
     let targets = env_parse(&cli.target_environment);
     let records = retrieve_from(&client, targets, cli.filter)?;
 
-    let results = consolidate_targets(records);
+    let results = ssr::consolidate_targets(records);
     println!("{:#?}", results);
 
     Ok(())
@@ -102,7 +64,7 @@ fn retrieve_from(
     client: &reqwest::blocking::RequestBuilder,
     targets: Vec<(String, String)>,
     pattern: Option<String>,
-) -> Result<Vec<(String, Vec<Ssr>)>> {
+) -> Result<Vec<(String, Vec<SsrRecord>)>> {
     let mut records = Vec::with_capacity(targets.len());
     let pattern = pattern.map(|val| val.to_lowercase());
 
@@ -118,22 +80,4 @@ fn retrieve_from(
         return Err(Error::NoRecordsToProcess);
     }
     Ok(records)
-}
-
-fn consolidate_targets(tasks: Vec<(String, Vec<Ssr>)>) -> Vec<SsrResult> {
-    let mut results: HashMap<String, SsrResult> = HashMap::new();
-
-    for target in tasks {
-        for result in target.1 {
-            if let Some(r) = results.get_mut(&result.key.clone()) {
-                r.update_url(&target.0, result.url);
-            } else {
-                let mut r = SsrResult::new(&result.name, &result.description, &result.key);
-                r.update_url(&target.0, result.url);
-                results.insert(result.key, r);
-            }
-        }
-    }
-
-    results.into_values().collect()
 }
